requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
cryptography==42.0.5 